Per recommendations on reddit and from other sources, I have implemented a new package structure.
