+++ src/cmsnbiclient/E7/__init__.py
@@ -1,0 +1,6 @@
+# E7 SUBPACKAGES IMPORT STATEMENTS
+from .create import Create
+from .delete import Delete
+from .query import Query
+from .update import Update
+# E7 SUBPACKAGES IMPORT STATEMENTS
