+++ src/cmsnbiclient/REST/__init__.py
@@ -1,0 +1,3 @@
+# REST SUBPACKAGES IMPORT STATEMENTS
+from .query import Query
+# REST SUBPACKAGES IMPORT STATEMENTS
