+++ src/cmsnbiclient/__init__.py
@@ -1,0 +1,24 @@
+# Versioning imports
+from .__version__ import (__version__,
+                          __url__,
+                          __title__,
+                          __author__,
+                          __author_email__,
+                          __license__,
+                          __description__)
+# Versioning imports
+
+# Package Dependency Imports
+import requests
+import xmltodict
+import pydash
+import json
+import os
+# Package Dependency Imports
+
+# Package Class Object imports
+from .client import Client
+from cmsnbiclient import E7
+from cmsnbiclient import REST
+# Package Class Object imports
+
