+++ src/cmsnbiclient/__version__.py
@@ -1,0 +1,9 @@
+__title__ = """cms_nbi_client"""
+__description__ = """CMS-NBI-Client is a python package in the works.
+                    The goal is to create a python package allowing for interaction with the NBI(Northbound Interfaces) of a CMS(Calix Management System) instance.
+                    This will allow for orchestration of the nodes managed by the CMS instances"""
+__url__ = """https://github.com/somenetworking/CMS-NBI-Client"""
+__version__ = """0.0.3"""
+__author__ = """somenetworking"""
+__author_email__ = """somenetworking@gmail.com"""
+__license__ = """GPL 3"""
